requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# LLM Providers
openai>=1.0.0
anthropic>=0.30.0

# Data Processing
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Async Support
asyncio-throttle>=1.0.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.0.0
structlog>=24.0.0

# Testing & Evaluation
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Visualization & Analysis
matplotlib>=3.8.0
seaborn>=0.13.0
tabulate>=0.9.0

# Progress & Logging
tqdm>=4.66.0
rich>=13.0.0

# Optional: For advanced metrics
scikit-learn>=1.4.0
scipy>=1.12.0
